refactor: API 타입 파일 경로 변경
