+++ src/features/challenge/hook/useInfiniteGroupChallengeVerifications.ts
@@ -1,7 +1,7 @@
 import { type InfiniteData, useInfiniteQuery } from '@tanstack/react-query'
 
 import { QUERY_KEYS } from '@shared/config/tanstack-query/query-keys'
-import { ApiResponse } from '@shared/lib/api/client-fetcher/type'
+import { ApiResponse } from '@shared/lib/api/type'
 
 import {
   getGroupChallengeParticipateList,
