+++ src/features/member/hooks/useInfiniteMemberAlarmList.ts
@@ -3,7 +3,7 @@
 import { type InfiniteData, useInfiniteQuery } from '@tanstack/react-query'
 
 import { QUERY_KEYS } from '@shared/config/tanstack-query/query-keys'
-import { ApiResponse } from '@shared/lib/api/client-fetcher/type'
+import { ApiResponse } from '@shared/lib/api/type'
 
 import { getMemberAlarmList, MemberAlarmList, MemberAlarmListParams } from '../api/get-alarm'
