+++ src/shared/lib/api/client-fetcher/client-fetcher.test.ts
@@ -3,8 +3,8 @@ import { beforeEach, describe, expect, it, MockedFunction, vi } from 'vitest'
 
 import type { EndpointType } from '../../../constants/endpoint/endpoint'
 import { HttpMethod, HttpStatusCode } from '../../../constants/http'
+import { ApiResponse, ErrorResponse } from '../type'
 import { fetchRequest } from './client-fetcher'
-import { ApiResponse, ErrorResponse } from './type'
 
 describe('fetchRequest 유틸', () => {
   let fetchMock: MockedFunction<typeof fetch>
