+++ src/shared/lib/api/server-fetcher/server-fetcher.ts
@@ -4,7 +4,7 @@ import { cookies } from 'next/headers'
 
 import { EndpointType } from '@shared/constants/endpoint/endpoint'
 
-import { ApiResponse, ErrorResponse, OptionsType } from '../client-fetcher/type'
+import { ApiResponse, ErrorResponse, OptionsType } from '../type'
 import { refreshServerAccessToken } from './server-reissue'
 
 const BASE_URL = 'https://leafresh.app'
