+++ src/shared/lib/api/type.ts
